Move starter pack splash image code to its own file
